content/media/video/test/test_wav_trunc.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Wave Media test: playback ends when file is truncated</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<audio id='v'
       onloadedmetadata='return startTest();'
       onended='return playbackEnded();'>
  <source type='audio/x-wav' src='r11025_u8_c1_trunc.wav'>
</audio>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test if the ended event works correctly.
var v = document.getElementById('v');
var endPassed = false;
var completed = false;

function startTest() {
  if (completed)
    return false;

  ok(v.duration > 1.7 && v.duration < 1.9, "Duration should be around 1.8: " + v.duration);

  v.play();
  return false;
}

function playbackEnded() {
  if (completed)
    return false

  completed = true;
  ok(v.currentTime >= 1.7 && v.currentTime <= 1.9,
     "Checking currentTime at end: " + v.currentTime);
  ok(v.ended, "Checking playback has ended");
  SimpleTest.finish();
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
